// workhub.client/src/generate-api/models/DeviceCategoryMinimalDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WorkHub.Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface DeviceCategoryMinimalDto
 */
export interface DeviceCategoryMinimalDto {
    /**
     * 
     * @type {number}
     * @memberof DeviceCategoryMinimalDto
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof DeviceCategoryMinimalDto
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceCategoryMinimalDto
     */
    description?: string | null;
}

/**
 * Check if a given object implements the DeviceCategoryMinimalDto interface.
 */
export function instanceOfDeviceCategoryMinimalDto(value: object): value is DeviceCategoryMinimalDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function DeviceCategoryMinimalDtoFromJSON(json: any): DeviceCategoryMinimalDto {
    return DeviceCategoryMinimalDtoFromJSONTyped(json, false);
}

export function DeviceCategoryMinimalDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): DeviceCategoryMinimalDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function DeviceCategoryMinimalDtoToJSON(json: any): DeviceCategoryMinimalDto {
    return DeviceCategoryMinimalDtoToJSONTyped(json, false);
}

export function DeviceCategoryMinimalDtoToJSONTyped(value?: DeviceCategoryMinimalDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'description': value['description'],
    };
}