// workhub.client/src/generate-api/models/DeviceMinimalDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WorkHub.Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { DeviceStatus } from './DeviceStatus';
import {
    DeviceStatusFromJSON,
    DeviceStatusFromJSONTyped,
    DeviceStatusToJSON,
    DeviceStatusToJSONTyped,
} from './DeviceStatus';

/**
 * 
 * @export
 * @interface DeviceMinimalDto
 */
export interface DeviceMinimalDto {
    /**
     * 
     * @type {number}
     * @memberof DeviceMinimalDto
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof DeviceMinimalDto
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof DeviceMinimalDto
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof DeviceMinimalDto
     */
    location?: string | null;
    /**
     * 
     * @type {DeviceStatus}
     * @memberof DeviceMinimalDto
     */
    status?: DeviceStatus;
}



/**
 * Check if a given object implements the DeviceMinimalDto interface.
 */
export function instanceOfDeviceMinimalDto(value: object): value is DeviceMinimalDto {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function DeviceMinimalDtoFromJSON(json: any): DeviceMinimalDto {
    return DeviceMinimalDtoFromJSONTyped(json, false);
}

export function DeviceMinimalDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): DeviceMinimalDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'location': json['location'] == null ? undefined : json['location'],
        'status': json['status'] == null ? undefined : DeviceStatusFromJSON(json['status']),
    };
}

export function DeviceMinimalDtoToJSON(json: any): DeviceMinimalDto {
    return DeviceMinimalDtoToJSONTyped(json, false);
}

export function DeviceMinimalDtoToJSONTyped(value?: DeviceMinimalDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'description': value['description'],
        'location': value['location'],
        'status': DeviceStatusToJSON(value['status']),
    };
}